moved logs previewers into logs-explorer section
