+++ studio/pages/project/[ref]/logs-explorer/sources/[source].tsx
@@ -3,18 +3,13 @@ import { NextPage } from 'next'
 import { useRouter } from 'next/router'
 import { observer } from 'mobx-react-lite'
 import { withAuth } from 'hooks'
-import { SettingsLayout } from 'components/layouts/'
-import { LOG_TYPE_LABEL_MAPPING, QueryType } from 'components/interfaces/Settings/Logs'
+import { LogsExplorerLayout } from 'components/layouts/'
+import { LogsTableName } from 'components/interfaces/Settings/Logs'
 import LogsPreviewer from 'components/interfaces/Settings/Logs/LogsPreviewer'
 
-/**
- * Placeholder page for logs previewers until we figure out where to slot them
- */
 export const LogPage: NextPage = () => {
   const router = useRouter()
-  const { ref, type } = router.query
-
-  const title = `Logs - ${LOG_TYPE_LABEL_MAPPING[type as keyof typeof LOG_TYPE_LABEL_MAPPING]}`
+  const { ref, source } = router.query
 
   useEffect(() => {
     if (filters.search_query !== s) {
@@ -98,13 +93,13 @@ export const LogPage: NextPage = () => {
   }
 
   return (
-    <SettingsLayout title={title}>
+    <LogsExplorerLayout subtitle={<span className="font-mono">{source}</span>}>
       <LogsPreviewer
         projectRef={ref as string}
-        queryType={type as QueryType}
         condensedLayout={true}
+        tableName={source as LogsTableName}
       />
-    </SettingsLayout>
+    </LogsExplorerLayout>
   )
 }
