+++ studio/tests/pages/projects/logs-query.test.js
@@ -44,7 +44,7 @@ import { useFlag } from 'hooks'
 useFlag.mockReturnValue(true)
 
 import { SWRConfig } from 'swr'
-import { LogsExplorerPage as Page } from 'pages/project/[ref]/logs-explorer/index'
+import { LogsExplorerPage as Page } from 'pages/project/[ref]/logs-explorer/query'
 const LogsExplorerPage = (props) => (
   <SWRConfig
     value={{
